requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.5
django-cors-headers==4.7.0
djangorestframework==3.15.2
mysqlclient==2.2.7
pillow==11.1.0
PyMySQL==1.1.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
